Correções finais

+++ PrimoHeranca/src/modelo/Propriedades.java
@@ -1,11 +1,9 @@
 package modelo;
 
-import java.math.BigInteger;
-
 public abstract class Propriedades
 {
     public String numero;
     public String mensagem;
     public String resposta;
-    public BigInteger num;
+    public int num;
 }
